Move ITelemetryManager to Core
